Add Method to Generate Object From Source Directories The method iterate_input iterates over the source directories making objects out of the folders that will be useful later. The objects contain the absolute path, the relative path, and whether the folder is at the supplied depth.
The depth option has the user specify how far into the source folder is the
level at which symlinks should be created for. With a directory structure like
music.flac/Artist/Album/, source_high would be music.flac and you would want to
create symlinks for the albums. That means --depth=2. If you linked to the
artist you could miss out on albums in different qualities.